requirements.txt
openai
requests
python-dotenv
gitpython